ndvi_app/requirements.txt
streamlit==1.24.0
rasterio
numpy
matplotlib
opencv-python
Pillow
streamlit-drawable-canvas
